Remove unused and PyQt unsupported codes
The `QtWidgets.QGraphicsSceneMouseEvent` class doesn't have `setModifiers()`
method in PyQt5, and cannot find a workaround for it.

So the only option to stick with PyQt5 for now was to drop the modifier
remapping, but judging from the event passing flow and running a few tests,
it looks like those remapping code didn't have any effect to the UI
function.

Should be safe to remove them.