feat: interpret menu function file relative to json path
this allows application code to construct an absolute
path to the shortcut menu json definition at runtime
(using importlib.resources or relative to __file__)
while the dynamically loaded python module path
('file' key) is assumed relative to the json path.
If an absolute filepath is given in the menu data dict,
that is used instead.